feat: renamed smartwatch app module for wearos app module
