+++ apps/wearos_app/.gitignore
