+++ apps/wearos_app/proguard-rules.pro
