+++ apps/wearos_app/src/main/AndroidManifest.xml
