+++ apps/wearos_app/src/main/ic_launcher-playstore.png
