+++ apps/wearos_app/src/main/kotlin/dev/marlonlom/apps/mocca/wearos/presentation/feats/main/WearAppContent.kt
@@ -3,12 +3,17 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package dev.marlonlom.apps.mocca.smartwatch.presentation.feats.main
+package dev.marlonlom.apps.mocca.wearos.presentation.feats.main
 
 import androidx.compose.runtime.Composable
 import androidx.wear.compose.foundation.lazy.rememberScalingLazyListState
-import dev.marlonlom.apps.mocca.smartwatch.presentation.theme.MoccaTheme
+import dev.marlonlom.apps.mocca.wearos.presentation.theme.MoccaTheme
 
+/**
+ * Wear app main content composable.
+ *
+ * @author marlonlom
+ */
 @Composable
 fun WearAppContent() {
   MoccaTheme {
