+++ apps/wearos_app/src/main/res/drawable/img_splash.webp
