+++ apps/wearos_app/src/main/res/drawable/splash_screen.xml
