+++ apps/wearos_app/src/main/res/font/comfortaa_regular.ttf
