+++ apps/wearos_app/src/main/res/mipmap-hdpi/ic_launcher.webp
