+++ apps/wearos_app/src/main/res/mipmap-xhdpi/ic_launcher.webp
