+++ apps/wearos_app/src/main/res/mipmap-xxhdpi/ic_launcher.webp
