+++ apps/wearos_app/src/main/res/mipmap-xxxhdpi/ic_launcher.webp
