+++ apps/wearos_app/src/main/res/values-es/strings.xml
