+++ apps/wearos_app/src/main/res/values/colors.xml
