+++ apps/wearos_app/src/main/res/values/ic_launcher_background.xml
