+++ apps/wearos_app/src/main/res/values/strings.xml
