+++ apps/wearos_app/src/main/res/values/styles.xml
